requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Flask-Migrate
python-dotenv
PyPDF2
openpyxl
Pillow
python-multipart
Werkzeug